refactor to get user dash input to work again
